pyproject.toml
[tool.poetry]
name = "daily-digest"
version = "0.1.0"
description = "Daily Digest PoC - Multi-team Slack summarizer using LangChain agents"
authors = ["Abhinav Prakash"]
readme = "README.md"
packages = [{include = "daily_digest", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
slack-sdk = "^3.21.3"
langchain = "^0.3.3"
langchain-openai = "^0.2.2"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"

[tool.poetry.scripts]
daily-digest = "daily_digest.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
